changed type of amount and balance from double to BigDecimal; added properties for creating payments and reworked ordering of transactions in T history; added global properties for initial payments created and delay between payment creations (just for initial ones)

+++ src/main/java/drdm/school/pia/manager/AccountManager.java
@@ -4,6 +4,8 @@ import drdm.school.pia.domain.entities.Account;
 import drdm.school.pia.domain.exceptions.PaymentValidationException;
 import drdm.school.pia.domain.entities.User;
 
+import java.math.BigDecimal;
+
 /**
  * Account manager
  * @author Michal Drda
@@ -21,7 +23,7 @@ public interface AccountManager {
      * @param account account which balance to be adjusted
      * @param valueOfChange valueOfChange provided as an addition or a substraction from the current account balance
      */
-    void updateBalance(Account account, double valueOfChange) throws PaymentValidationException;
+    void updateBalance(Account account, BigDecimal valueOfChange) throws PaymentValidationException;
 
     /**
      * Finds an account by given user name
